requirements.txt
asn1crypto==0.24.0
astroid==2.0.4
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.4.2
Django==2.1.3
django-cors-headers==2.4.0
django-rest-knox==3.4.0
django-rest-swagger==2.2.0
djangorestframework==3.9.0
idna==2.7
isort==4.3.4
itypes==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.1.0
mccabe==0.6.1
openapi-codec==1.3.2
pycparser==2.19
pylint==2.1.1
pylint-django==2.0.2
pylint-plugin-utils==0.4
pyOpenSSL==18.0.0
pytz==2018.7
requests==2.20.1
simplejson==3.16.0
six==1.11.0
uritemplate==3.0.0
urllib3==1.24.1
wrapt==1.10.11
